refactor: separate modules for component and service
