+++ package.json
@@ -7,6 +7,8 @@
     "build": "ng build",
     "test": "ng test",
     "lint": "ng lint",
+    "link-lib": "cd dist/coreui-icons-angular/ && npm link",
+    "publish-lib": "cd dist/coreui-icons-angular/ && npm publish --tag next --dry-run",
     "e2e": "ng e2e"
   },
   "private": true,
@@ -46,5 +48,16 @@
     "tslint": "~6.1.0",
     "typescript": "~3.7.5",
     "@coreui/icons": "^2.0.0-beta.5"
-  }
+  },
+  "keywords": [
+    "coreui",
+    "coreui-icons",
+    "coreui-angular",
+    "icons",
+    "svg",
+    "svg-icons",
+    "layout",
+    "component",
+    "angular"
+  ]
 }
