+++ projects/coreui/icons-angular/package.json
@@ -7,5 +7,16 @@
   },
   "dependencies": {
     "tslib": "^1.10.0"
-  }
+  },
+  "keywords": [
+    "coreui",
+    "coreui-icons",
+    "coreui-angular",
+    "icons",
+    "svg",
+    "svg-icons",
+    "layout",
+    "component",
+    "angular"
+  ]
 }
