+++ projects/coreui/icons-angular/src/lib/icon-set/icon-set.service.ts
@@ -7,7 +7,7 @@ export interface IIconSet {
 @Injectable({
   providedIn: 'root'
 })
-export class IconService {
+export class IconSetService {
 
   // tslint:disable-next-line:variable-name
   private _icons: IIconSet = {};
