+++ projects/coreui/icons-angular/src/lib/icon-set/index.ts
