+++ projects/coreui/icons-angular/src/lib/icon-set/public_api.ts
@@ -1,0 +1,2 @@
+export { IconSetService, IIconSet } from './icon-set.service';
+export { IconSetModule } from './icon-set.module';
