+++ projects/coreui/icons-angular/src/lib/icon/icon.component.html
@@ -1,5 +1,6 @@
 <img *ngIf="src"
      [cHtmlAttr]="attributes"
+     [attr.alt]="title"
      [src]="src"
 />
 <svg  *ngIf="(!src) && (!use)"
@@ -7,18 +8,20 @@
       [attr.width]="width"
       [attr.height]="height || width"
       [attr.viewBox]="viewBox"
-      [attr.title]="titleCode"
       [innerHtml]="iconCode"
       [attr.class]="computedClasses"
+      [cHtmlAttr]="attributes"
       role="img"
       pointer-events="none"
 >
+  {{titleCode}}
 </svg>
 <svg *ngIf="use"
      xmlns="http://www.w3.org/2000/svg"
      [attr.width]="width"
      [attr.height]="height || width"
      [attr.class]="computedClasses"
+     [cHtmlAttr]="attributes"
      role="img"
      pointer-events="none"
 >
