+++ projects/coreui/icons-angular/src/lib/icon/icon.component.scss
