+++ projects/coreui/icons-angular/src/lib/icon/icon.module.ts
@@ -1,0 +1,20 @@
+import {NgModule} from '@angular/core';
+import { CommonModule } from '@angular/common';
+
+import {IconComponent} from './icon.component';
+import { HtmlAttributesDirective } from '../shared/html-attr.directive';
+
+@NgModule({
+  declarations: [
+    IconComponent,
+    HtmlAttributesDirective
+  ],
+  imports: [
+    CommonModule,
+  ],
+  exports: [
+    IconComponent,
+    HtmlAttributesDirective
+  ],
+})
+export class IconModule {}
