+++ projects/coreui/icons-angular/src/lib/icon/public_api.ts
@@ -1,3 +1,2 @@
 export { IconComponent } from './icon.component';
-export { IconService, IIconSet } from './icon.service';
 export { IconModule } from './icon.module';
