+++ projects/coreui/icons-angular/src/lib/shared/html-attr.directive.spec.ts
